Start to cleanup debianization
Fixes many lintian complaints, mostly by deleting unused template files,
adding a manpage, not marking images as executable, and removing the
unused font file (it is currently embedded in the binary, so no need to
distribute the file, though depending on fonts-lato would be even
nicer).

Also, update to use the pkg-config options supported by Projucer instead
of doing pkg-config stuff manually in the root Makefile and add new
source files to the Projucer project. This allows Projucer to build the
project again.